requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deep-translator==1.11.4
defusedxml==0.7.1
Django==5.2.8
django-countries==8.1.1
dj-database-url==2.2.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
gunicorn==22.0.0
httplib2==0.31.0
idna==3.11
joblib==1.5.2
nltk==3.9.2
oauthlib==3.3.1
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
python3-openid==3.2.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
soupsieve==2.8
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.7.0
asgiref
beautifulsoup4
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
deep-translator
defusedxml
dj-database-url
Django
django-countries
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gunicorn
httplib2
idna
joblib
nltk
oauthlib
pillow
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyparsing
python-dotenv
python3-openid
regex
requests
requests-oauthlib
rsa
social-auth-app-django
social-auth-core
soupsieve
sqlparse
textblob
tqdm
typing_extensions
tzdata
uritemplate
urllib3
whitenoise
